requirements.txt
Flask
pdfplumber
requests
beautifulsoup4
python-docx
